add: checking access rights for editing or deleting objects and comments.

+++ src/main/java/com/example/ratingsystem/service/CommentService.java
@@ -1,8 +1,10 @@
 package com.example.ratingsystem.service;
 
 import com.example.ratingsystem.domain.entities.Comment;
+import com.example.ratingsystem.domain.entities.User;
 import com.example.ratingsystem.repository.CommentRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -12,6 +14,7 @@ import java.util.UUID;
 @Service
 @RequiredArgsConstructor
 public class CommentService {
+    private final AuthService authService;
     private final CommentRepository commentRepository;
 
     public Comment create(Comment comment) {
@@ -19,11 +22,16 @@ public class CommentService {
     }
 
     public Comment update(Comment comment) {
+        validateAuthorAccess(comment.getAuthor());
         return commentRepository.save(comment);
     }
 
     public void delete(UUID id) {
-        commentRepository.deleteById(id);
+        commentRepository.findById(id)
+                .ifPresent(comment -> {
+                    validateAuthorAccess(comment.getAuthor());
+                    commentRepository.deleteById(id);
+                });
     }
 
     public Optional<Comment> get(UUID id) {
@@ -32,5 +40,11 @@ public class CommentService {
 
     public List<Comment> getBySellerId(UUID sellerId) {
         return commentRepository.findBySellerId(sellerId);
+    }
+
+    private void validateAuthorAccess(User author) {
+        if (author == null || !authService.validateAuthenticatedUser(author.getId())) {
+            throw new AccessDeniedException("Only author can manage their comments");
+        }
     }
 }
