+++ src/main/java/com/example/ratingsystem/service/GameObjectService.java
@@ -1,8 +1,10 @@
 package com.example.ratingsystem.service;
 
 import com.example.ratingsystem.domain.entities.GameObject;
+import com.example.ratingsystem.domain.entities.User;
 import com.example.ratingsystem.repository.GameObjectRepository;
-import lombok.AllArgsConstructor;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -10,20 +12,26 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Service
-@AllArgsConstructor
+@RequiredArgsConstructor
 public class GameObjectService {
-    private GameObjectRepository gameObjectRepository;
+    private final AuthService authService;
+    private final GameObjectRepository gameObjectRepository;
 
-    public GameObject create(GameObject dto) {
-        return gameObjectRepository.save(dto);
+    public GameObject create(GameObject object) {
+        return gameObjectRepository.save(object);
     }
 
-    public GameObject update(GameObject dto) {
-        return gameObjectRepository.save(dto);
+    public GameObject update(GameObject object) {
+        validateSellerAccess(object.getUser());
+        return gameObjectRepository.save(object);
     }
 
     public void delete(UUID id) {
-        gameObjectRepository.deleteById(id);
+        gameObjectRepository.findById(id)
+                .ifPresent(object -> {
+                    validateSellerAccess(object.getUser());
+                    gameObjectRepository.deleteById(id);
+                });
     }
 
     public Optional<GameObject> get(UUID id) {
@@ -40,5 +48,11 @@ public class GameObjectService {
 
     public List<GameObject> getAll() {
         return gameObjectRepository.findAll();
+    }
+
+    private void validateSellerAccess(User objectOwner) {
+        if (!authService.validateAuthenticatedUser(objectOwner.getId())) {
+            throw new AccessDeniedException("Only seller can manage his objects.");
+        }
     }
 }
